requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiounittest==1.5.0
arrow==1.4.0
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
binaryornot==0.4.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
Django==5.2.8
django-tailwind==4.4.1
fake-useragent==2.2.0
frozenlist==1.8.0
howlongtobeatpy==1.0.18
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
propcache==0.4.1
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==2.0.1
yarl==1.22.0
